Revert "提交代码 升级JDK与springboot升级"
This reverts commit 1ecdbd66965776193c60996d4a6fc29bea4197d6.

+++ sweet-home/src/main/java/com/gitdatsanvich/sweethome/model/dto/AccessDTO.java
@@ -15,7 +15,7 @@ import java.io.Serializable;
 @AllArgsConstructor
 public class AccessDTO implements Serializable {
     private String ip;
-    private Long accessTime;
+    private Integer accessTime;
     private Boolean accessAble;
-    private Long totalAccessTime;
+    private Integer totalAccessTime;
 }
